Moved setInjectComponent and injectComponent to uiinjection.js and updated references.
Updated OnScreenKeyboard to reference new location for injectComponent.

moved ContextMenuType and registerPluginContextMenuItem methods to
uiinjection.js

Removed addContextMenuItem action.
