+++ src/plugins/OnScreenKeyboard/index.js
@@ -1,7 +1,7 @@
 'use strict'
 
 import OnScreenKeyboard from './OnScreenKeyboard.vue'
-import { injectComponent } from '..'
+import { injectComponent } from '../../store/uiinjection'
 
 injectComponent('on-screen-keyboard', OnScreenKeyboard)
