+++ src/store/uiinjection.js
@@ -1,4 +1,3 @@
-
 export default  {
     namespaced : true,
     state: {
@@ -6,14 +5,36 @@ export default  {
 			jobFileList: []
 		}
     },
-    actions: {
-		addContextMenuItem({ commit }, contextMenuItem) {
-			commit('addContextMenuItem', contextMenuItem);
-		}
-    },
     mutations : {
 		addContextMenuItem(state, payload) {
 			state.contextMenuItems[payload.menu].push(payload.item);
 		}
     }
+}
+
+import store from '.'
+
+export const ContextMenuType = {
+	JobFileList: 'jobFileList'
+}
+
+export function registerPluginContextMenuItem(name, path, icon, action, contextMenuType) {
+    store.commit('uiinjection/addContextMenuItem', {
+		menu: contextMenuType,
+		item: {
+			name: name,
+			path: path,
+			icon: icon,
+			action: action
+		}
+	});
+}
+
+let injectComponentFn
+export function setInjectComponent(fn) {
+	injectComponentFn = fn;
+}
+
+export function injectComponent(name, component) {
+	injectComponentFn(name, component)
 }
